double change

+++ tuple.c
@@ -6,13 +6,13 @@
 /*   By: zcanales <zcanales@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 09:52:37 by zcanales          #+#    #+#             */
-/*   Updated: 2022/04/20 11:04:39 by zcanales         ###   ########.fr       */
+/*   Updated: 2022/04/20 13:51:16 by zcanales         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 
-void print_tuple(float *tpl)
+void print_tuple(double *tpl)
 {
 	int i = -1;
 	while (++i < 4)
@@ -20,11 +20,11 @@ void print_tuple(float *tpl)
 	printf("\n");
 }
 
-float *create_tuple(float x, float y, float z, float w)
+double *create_tuple(double x, double y, double z, double w)
 {
-	float *tuple;
+	double *tuple;
 
-	tuple = (float *)malloc(sizeof(float) * 4);
+	tuple = (double *)malloc(sizeof(double) * 4);
 	if (!tuple)
 		return (NULL);
 	tuple[0] = x;
@@ -34,12 +34,12 @@ float *create_tuple(float x, float y, float z, float w)
 	return (tuple);
 }
 
-float *add_tuple(float *tpl1, float* tpl2)
+double *add_tuple(double *tpl1, double* tpl2)
 {
-	float *add_tpl;
+	double *add_tpl;
 	int i;
 
-	add_tpl = (float *)malloc(sizeof(float) * 4);
+	add_tpl = (double *)malloc(sizeof(double) * 4);
 	if (!add_tpl)
 		return (NULL);
 	i = -1;
@@ -48,12 +48,12 @@ float *add_tuple(float *tpl1, float* tpl2)
 	return (add_tpl);
 }
 
-float *sub_tuple(float *tpl1, float* tpl2)
+double *sub_tuple(double *tpl1, double* tpl2)
 {
-	float *sub_tpl;
+	double *sub_tpl;
 	int i;
 
-	sub_tpl = (float *)malloc(sizeof(float) * 4);
+	sub_tpl = (double *)malloc(sizeof(double) * 4);
 	if (!sub_tpl)
 		return (NULL);
 	i = -1;
@@ -64,16 +64,16 @@ float *sub_tuple(float *tpl1, float* tpl2)
 
 int main()
 {
-	float *t1 = create_tuple(1, 2, 3, 1);
-	float *t2 = create_tuple(2, 3, 4, 1);
-	float *t3 = add_tuple(t1, t2);
-	float *t4 = scalar_mul_vec(t1, 2);
-	float *t5 = scalar_div_vec(t1, 2);
+	double *t1 = create_tuple(1, 2, 3, 1);
+	double *t2 = create_tuple(2, 3, 4, 1);
+	double *t3 = add_tuple(t1, t2);
+	double *t4 = scalar_mul_vec(t1, 2);
+	double *t5 = scalar_div_vec(t1, 2);
 	printf("dot product = %f\n",  dot_product_vec(t1, t2));
-	float *t6 = cross_product_vec(t1, t2);
-	float *t7 = neg_vec(t1);
+	double *t6 = cross_product_vec(t1, t2);
+	double *t7 = neg_vec(t1);
 	printf("magnitude = %f\n", magnitude_vec(t1));
-	float *t8 = normalization_vec(t1, magnitude_vec(t1));
+	double *t8 = normalization_vec(t1, magnitude_vec(t1));
 
 	print_tuple(t4);
 	print_tuple(t5);
