+++ vector_operations2.c
@@ -6,19 +6,19 @@
 /*   By: zcanales <zcanales@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 10:37:55 by zcanales          #+#    #+#             */
-/*   Updated: 2022/04/20 11:04:29 by zcanales         ###   ########.fr       */
+/*   Updated: 2022/04/20 13:51:15 by zcanales         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 #include "math.h"
 
-float *neg_vec(float *tpl)
+double *neg_vec(double *tpl)
 {
-	float *result;
+	double *result;
 	int i;
 
-	result = (float *)malloc(sizeof(float) * 4);
+	result = (double *)malloc(sizeof(double) * 4);
 	i = -1;
 	while (++i < 3)
 		result[i] = tpl[i] * -1;
@@ -26,10 +26,10 @@ float *neg_vec(float *tpl)
 	return (result);
 }
 
-float magnitude_vec(float *tpl)
+double magnitude_vec(double *tpl)
 {
 	int i;
-	float result;
+	double result;
 
 	i = -1;
 	result = 0;
@@ -39,12 +39,12 @@ float magnitude_vec(float *tpl)
 
 }
 
-float *normalization_vec(float *tpl, float magnitude)
+double *normalization_vec(double *tpl, double magnitude)
 {
-	float *result;
+	double *result;
 	int i;
 
-	result = (float *)malloc(sizeof(float) * 4);
+	result = (double *)malloc(sizeof(double) * 4);
 	i = -1;
 	while (++i < 3)
 		result[i] = tpl[i] / magnitude;
